# api/swagger/swagger.json
{
  "info": {
    "title": "Discovery Service",
    "description": "This service helps you discovery other services",
    "version": "1.0.0"
  },
  "swagger": "2.0",
  "paths": {
    "/constants": {
      "get": {
        "description": "Get Service Constants",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Constants",
            "type": "object"
          }
        }
      }
    },
    "/health": {
      "get": {
        "description": "Get Health of Service",
        "tags": [
          "health"
        ],
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Health",
            "type": "object",
            "schema": {
              "$ref": "#/definitions/Health"
            }
          }
        }
      }
    },
    "/services": {
      "get": {
        "description": "Get Health of Service",
        "tags": [
          "services"
        ],
        "operationId": "getServices",
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "types",
            "description": "Types of Services",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "stageFilter",
            "description": "Stage Filter",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "regionFilter",
            "description": "Region Filter",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "page",
            "description": "Page Number",
            "in": "query",
            "required": false,
            "type": "number"
          },
          {
            "name": "size",
            "description": "Page size",
            "in": "query",
            "type": "number"
          }
        ],
        "responses": {
          "200": {
            "description": "PageResponse",
            "schema": {
              "$ref": "#/definitions/PageResponse"
            }
          }
        }
      }
    },
    "/services/_types": {
      "get": {
        "description": "Get Unique Service Types",
        "tags": [
          "services"
        ],
        "operationId": "uniqueServiceTypes",
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "ServiceTypes",
            "type": "array",
            "items": {
              "schema": {
                "$ref": "#/definitions/ServiceType"
              }
            }
          }
        }
      }
    },
    "/services/_count": {
      "get": {
        "description": "Get Count of Services",
        "tags": [
          "services"
        ],
        "operationId": "count",
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "stageFilter",
            "description": "Stage Filter",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "regionFilter",
            "description": "Region Filter",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Count",
            "schema": {
              "$ref": "#/definitions/Count"
            }
          }
        }
      }
    },
    "/services/{id}": {
      "get": {
        "description": "Get Service By Id",
        "tags": [
          "services"
        ],
        "operationId": "getServiceById",
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "description": "Service Id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "ServiceDescriptor",
            "schema": {
              "$ref": "#/definitions/ServiceDescriptor"
            }
          }
        }
      },
      "delete": {
        "description": "Get Service By Id",
        "operationId": "deleteServiceById",
        "tags": [
          "services"
        ],
        "produces": [
          "application/json"
        ],
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "description": "Service Id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "ServiceDescriptor",
            "schema": {
              "$ref": "#/definitions/ServiceDescriptor"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "Error": {
      "type": "object",
      "required": [
        "errorMessage"
      ],
      "properties": {
        "errorMessage": {
          "type": "string"
        }
      }
    },
    "Health": {
      "type": "object",
      "required": [
        "cpuPercentUsage",
        "totalMemPercentageUsage",
        "loadAvg"
      ],
      "properties": {
        "cpuPercentUsage": {
          "type": "number"
        },
        "totalMemPercentageUsage": {
          "type": "number"
        },
        "loadAvg": {
          "type": "number"
        }
      }
    },
    "ServiceType": {
      "type": "object",
      "required": [
        "type"
      ],
      "properties": {
        "type": {
          "type": "string"
        }
      }
    },
    "PageDescriptor": {
      "type": "object",
      "required": [
        "page",
        "size",
        "total"
      ],
      "properties": {
        "page": {
          "type": "integer",
          "format": "int64"
        },
        "size": {
          "type": "integer",
          "format": "int64"
        },
        "total": {
          "type": "integer",
          "format": "int64"
        }
      }
    },
    "PageResponse": {
      "type": "object",
      "required": [
        "page",
        "elements"
      ],
      "properties": {
        "page": {
          "type": "object",
          "schema": {
            "$ref": "#/definitions/PageDescriptor"
          }
        },
        "elements": {
          "type": "array",
          "$ref": "#/definitions/ServiceDescriptor"
        }
      }
    },
    "Count": {
      "type": "object",
      "required": [
        "count"
      ],
      "properties": {
        "count": {
          "type": "integer",
          "format": "int64"
        }
      }
    },
    "ServiceDescriptor": {
      "type": "object",
      "required": [
        "username",
        "version",
        "type",
        "healthCheckRoute",
        "schemaRoute",
        "docsRoute",
        "status",
        "region",
        "stage"
      ],
      "properties": {
        "id": {
          "type": "string"
        },
        "version": {
          "type": "string"
        },
        "type": {
          "type": "string"
        },
        "healthCheckRoute": {
          "type": "string"
        },
        "schemaRoute": {
          "type": "string"
        },
        "docsRoute": {
          "type": "string"
        },
        "timestamp": {
          "type": "number",
          "format": "date"
        },
        "status": {
          "type": "string"
        },
        "region": {
          "type": "string"
        },
        "stage": {
          "type": "string"
        },
        "rtimes": {
          "type": "array",
          "items": {
            "type": "number"
          }
        },
        "avgTime": {
          "type": "number"
        }
      }
    }
  },
  "responses": {},
  "parameters": {},
  "securityDefinitions": {},
  "tags": []
}